Revert "Merge branch 'dev' into master"
This reverts commit 3611bd83ff53290b3437c5f39d3050e27ae76395, reversing
changes made to bc9199148739afc7f9e3d0fc7d520b31981c58db.

+++ server-plugin/Code/Systems/Blockers/BhopBlocker.cpp
@@ -17,6 +17,8 @@ limitations under the License.
 
 #include <cmath>
 
+#include "Interfaces/InterfacesProxy.h"
+
 #include "Systems/Logger.h"
 
 BhopBlocker::BhopBlocker () :
@@ -130,7 +132,7 @@ PlayerRunCommandRet BhopBlocker::RT_PlayerRunCommandCallback ( PlayerHandler::co
 	float const tick_interval ( SourceSdk::InterfacesProxy::Call_GetTickInterval () );
 	__assume ( tick_interval > 0.0f && tick_interval < 1.0f );
 
-	int const blocking_delay_ticks ( ( int )( floorf ( blocking_delay_seconds / tick_interval ) ) );
+	int const blocking_delay_ticks ( floorf ( blocking_delay_seconds / tick_interval ) );
 
 	int const gtick ( Helpers::GetGameTickCount () );
